XIVY-11544 Add ElementId and SelectedElement with app and pmv information
Added a new module vscode-base that defines the interface between the process editor and inscription view extension. It defines the ProcessEditorExtension, ProcessEditorConnector and SelectedElement and ElementId interfaces.

The process editor extension reads the app and pmv provided as args in the root model. It enhances the selection of the standard GLSP client with those information.

The inscription view extensions listen on selected element changes and sets the information to the inscription web view.